Rename PanelSideBarMenuItem.ts to PanelItem.ts and removed TMenuItem
