+++ src/index.ts
@@ -4,7 +4,7 @@ export * from "./lib/Layout/AuthenticationLayout";
 // Panel sidebar layout
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBarLayout";
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext";
-export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelSideBarMenuItem";
+export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelItem";
 
 // Sidebar layout
 export * from "./lib/Layout/SideBarLayout/SideBarLayout";
