+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/PanelSidebar.tsx
@@ -2,7 +2,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import classNames from "classnames";
 import { Button } from "reactstrap";
 import { usePanelSideBarContext } from "./Context/PanelSideBarContext";
-import { PanelItem } from "./Definitions/PanelSideBarMenuItem";
+import { PanelItem } from "./Definitions/PanelItem";
 import { PanelSideBarItem } from "./PanelSideBarItem";
 
 interface PanelSidebarProps {
